add orthographic camera
the camera will now respond to the window resizing, ensuring that our
desired world coordinate space is always rendered with the correct
aspect ratio.

because this uses some recent changes to amethyst, a specific revision
of amethyst is now targeted. when the next release lands, this change
will be reverted