standardize vec2 format
the vec2 format was standardized on the `nalgebra::Vector2` type. most
math is updated to take advantage of the nalgebra library. a `vector`
module was added to make vector operations easier. this module, along
with the `util` module, has a prelude that can be glob imported for
convenience

the motion system was overhauled to use a new bookkeeping component to
track the entity's previous position, current position, and current
velocity. this should make it easier to prevent collisions with world
geometry: if a collision is detected, the position update of the entity is
simply rolled back.

lastly, because this is a fairly comprehensive update, the version of
`lase` is now bumped to `0.4.0`.